src/main/java/is/galia/util/IOUtils.java
/*
 * Copyright © 2024 Baird Creek Software LLC
 *
 * Licensed under the PolyForm Noncommercial License, version 1.0.0;
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at:
 *
 *     https://polyformproject.org/licenses/noncommercial/1.0.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package is.galia.util;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import javax.imageio.stream.ImageInputStream;
import java.io.IOException;
import java.io.OutputStream;

public final class IOUtils {

    private static final Logger LOGGER = LoggerFactory.getLogger(IOUtils.class);

    /**
     * Invokes {@link AutoCloseable#close()} on the given instance quietly but
     * not silently (exceptions are logged).
     */
    public static void closeQuietly(AutoCloseable closeable) {
        if (closeable != null) {
            try {
                closeable.close();
            } catch (Exception e) {
                LOGGER.warn(e.getMessage());
            }
        }
    }

    public static void consume(ImageInputStream inputStream) throws IOException {
        try (OutputStream os = OutputStream.nullOutputStream()) {
            transfer(inputStream, os);
        }
    }

    public static void transfer(ImageInputStream inputStream,
                                OutputStream outputStream) throws IOException {
        byte[] buffer = new byte[16384];
        int len;
        while ((len = inputStream.read(buffer)) != -1) {
            outputStream.write(buffer, 0, len);
        }
    }

    private IOUtils() {}

}
